Calc. Zonal Statistics of a Raster using Python
https://www.youtube.com/watch?v=VIr-pejky6E

his tutorial explains the basics of calculating the zonal statistics such as average, max, min values corresponding to a defined spatial area of a raster using a polygon vector layer using Python, with the aid of GeoPandas, rasterio and rasterstats libraries. Additionally, you will learn how to plot rasters and vectors on the same layer, how to use matplotlib subplots to plot maps and histograms together in the same canvas, as well as to extract the summary of the data into a Pandas DataFrame using the Pandas library.